# api/openapi.json
{
  "components": {
    "schemas": {
...
      "CreatePurchaseDto": {
        "properties": {
          "programCode": {
            "maxLength": 32,
            "minLength": 1,
            "type": "string"
          },
          "quoteCurrency": {
            "maxLength": 3,
            "minLength": 3,
            "type": "string"
          },
          "tokenAmount": {
            "minimum": 1,
            "type": "number"
          }
        },
        "required": [
          "programCode",
          "tokenAmount"
        ],
        "type": "object"
      },
      "CreateRedemptionDto": {
        "properties": {
          "amount": {
            "minimum": 1,
            "type": "number"
          },
          "note": {
            "maxLength": 500,
            "type": "string"
          },
          "programCode": {
            "maxLength": 32,
            "minLength": 1,
            "type": "string"
          }
        },
        "required": [
          "programCode",
          "amount"
        ],
        "type": "object"
      },
      "CreateWaitlistDto": {
        "properties": {
          "consent": {
            "description": "Must be exactly true.",
            "enum": [
              true
            ],
            "type": "boolean"
          },
          "email": {
            "type": "string"
          },
          "fullName": {
            "maxLength": 200,
            "type": "string"
          },
          "interest": {
            "maxLength": 500,
            "type": "string"
          },
          "investorType": {
            "enum": [
              "institution",
              "investor",
              "partner",
              "other"
            ],
            "type": "string"
          },
          "locale": {
            "maxLength": 10,
            "type": "string"
          },
          "organization": {
            "maxLength": 200,
            "type": "string"
          }
        },
        "required": [
          "consent",
          "fullName",
          "email",
          "investorType"
        ],
        "type": "object"
      },
...
      "LinkWalletDto": {
        "properties": {
          "address": {
            "pattern": "/^0x[a-fA-F0-9]{40}$/",
            "type": "string"
          },
          "chainId": {
            "enum": [
              11155111
            ],
            "type": "number"
          }
        },
        "required": [
          "address"
        ],
        "type": "object"
      },
...
      "RejectPurchaseDto": {
        "properties": {
          "reason": {
            "maxLength": 2000,
            "type": "string"
          }
        },
        "required": [
          "reason"
        ],
        "type": "object"
      },
      "RejectRedemptionDto": {
        "properties": {
          "reason": {
            "maxLength": 2000,
            "type": "string"
          }
        },
        "required": [
          "reason"
        ],
        "type": "object"
      },
...
      "SettlePurchaseDto": {
        "properties": {
          "txRef": {
            "maxLength": 200,
            "type": "string"
          }
        },
        "type": "object"
      },
      "SettleRedemptionDto": {
        "properties": {
          "txRef": {
            "maxLength": 200,
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "WaitlistCreated": {
        "properties": {
          "id": {
            "type": "string"
          },
          "ok": {
            "enum": [
              true
            ],
            "type": "boolean"
          }
        },
        "required": [
          "ok",
          "id"
        ],
        "type": "object"
      },
      "WaitlistEntryResponse": {
        "properties": {
          "consent": {
            "type": "boolean"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "email": {
            "description": "PII.",
            "type": "string"
          },
          "fullName": {
            "description": "PII.",
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "interest": {
            "nullable": true,
            "type": "string"
          },
          "investorType": {
            "enum": [
              "institution",
              "investor",
              "partner",
              "other"
            ],
            "type": "string"
          },
          "locale": {
            "enum": [
              "en",
              "es",
              "it"
            ],
            "type": "string"
          },
          "organization": {
            "description": "PII.",
            "nullable": true,
            "type": "string"
          }
        },
        "required": [
          "id",
          "fullName",
          "email",
          "investorType",
          "consent",
          "organization",
          "interest",
          "locale",
          "createdAt"
        ],
        "type": "object"
      }
    },
    "securitySchemes": {
      "bearer": {
        "bearerFormat": "JWT",
        "description": "Admin domain (typ=admin, JWT_SECRET) on admin routes; investor domain (typ=investor, INVESTOR_JWT_SECRET) on /investor/me and /investor/status. The two are disjoint and neither is accepted by the other.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "ReserveChain API.\n\nIN DEVELOPMENT. No tokens are offered or sold.\n\nProof-of-Reserves, redemption, wallet linking and token purchase are\npublished as a contract only. Every operation tagged `inactive` returns\n501 and is marked deprecated so a generated client warns at the call site.\nThey require written authorization before they do anything.\n\nNo example values are supplied anywhere in this document. Every response\nfield is described by type and constraint only.",
    "title": "ReserveChain API",
    "version": "0.1.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/api/purchase/intents": {
      "get": {
        "deprecated": true,
        "operationId": "Purchase_list",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "take",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "501": {
            "description": "INACTIVE (P14). Gated by PURCHASE_ENABLED; with the flag on the service still refuses. Requires written authorization before it does anything."
          }
        },
        "tags": [
          "inactive"
        ]
      },
      "post": {
        "deprecated": true,
        "operationId": "Purchase_create",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePurchaseDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "501": {
            "description": "INACTIVE (P14). Gated by PURCHASE_ENABLED; with the flag on the service still refuses. Requires written authorization before it does anything."
          }
        },
        "tags": [
          "inactive"
        ]
      }
    },
...
    "/api/purchase/intents/{id}/approve": {
      "post": {
        "deprecated": true,
        "operationId": "Purchase_approve",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "501": {
            "description": "INACTIVE (P14). Gated by PURCHASE_ENABLED; with the flag on the service still refuses. Requires written authorization before it does anything."
          }
        },
        "tags": [
          "inactive"
        ]
      }
    },
    "/api/purchase/intents/{id}/reject": {
      "post": {
        "deprecated": true,
        "operationId": "Purchase_reject",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectPurchaseDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "501": {
            "description": "INACTIVE (P14). Gated by PURCHASE_ENABLED; with the flag on the service still refuses. Requires written authorization before it does anything."
          }
        },
        "tags": [
          "inactive"
        ]
      }
    },
    "/api/purchase/intents/{id}/settle": {
      "post": {
        "deprecated": true,
        "operationId": "Purchase_settle",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SettlePurchaseDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "501": {
            "description": "INACTIVE (P14). Gated by PURCHASE_ENABLED; with the flag on the service still refuses. Requires written authorization before it does anything."
          }
        },
        "tags": [
          "inactive"
        ]
      }
    },
...
    "/api/redemption/requests": {
      "get": {
        "deprecated": true,
        "operationId": "Redemption_list",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "take",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "501": {
            "description": "INACTIVE (P12). Gated by REDEMPTION_ENABLED; with the flag on the service still refuses. Requires written authorization before it does anything."
          }
        },
        "tags": [
          "inactive"
        ]
      },
      "post": {
        "deprecated": true,
        "operationId": "Redemption_create",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRedemptionDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "501": {
            "description": "INACTIVE (P12). Gated by REDEMPTION_ENABLED; with the flag on the service still refuses. Requires written authorization before it does anything."
          }
        },
        "tags": [
          "inactive"
        ]
      }
    },
...
    "/api/redemption/requests/{id}/approve": {
      "post": {
        "deprecated": true,
        "operationId": "Redemption_approve",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "501": {
            "description": "INACTIVE (P12). Gated by REDEMPTION_ENABLED; with the flag on the service still refuses. Requires written authorization before it does anything."
          }
        },
        "tags": [
          "inactive"
        ]
      }
    },
    "/api/redemption/requests/{id}/reject": {
      "post": {
        "deprecated": true,
        "operationId": "Redemption_reject",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectRedemptionDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "501": {
            "description": "INACTIVE (P12). Gated by REDEMPTION_ENABLED; with the flag on the service still refuses. Requires written authorization before it does anything."
          }
        },
        "tags": [
          "inactive"
        ]
      }
    },
    "/api/redemption/requests/{id}/settle": {
      "post": {
        "deprecated": true,
        "operationId": "Redemption_settle",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SettleRedemptionDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "501": {
            "description": "INACTIVE (P12). Gated by REDEMPTION_ENABLED; with the flag on the service still refuses. Requires written authorization before it does anything."
          }
        },
        "tags": [
          "inactive"
        ]
      }
    },
...
    "/api/waitlist": {
      "get": {
        "operationId": "Waitlist_list",
        "parameters": [
          {
            "in": "query",
            "name": "take",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/WaitlistEntryResponse"
                  },
                  "type": "array"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "waitlist"
        ]
      },
      "post": {
        "operationId": "Waitlist_create",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWaitlistDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WaitlistCreated"
                }
              }
            },
            "description": ""
          },
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WaitlistCreated"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "waitlist"
        ]
      }
    },
...
    "/api/wallet": {
      "get": {
        "deprecated": true,
        "operationId": "Wallet_list",
        "parameters": [
          {
            "in": "query",
            "name": "take",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "501": {
            "description": "INACTIVE (P13). Gated by WALLET_ENABLED; with the flag on the service still refuses. Requires written authorization before it does anything."
          }
        },
        "tags": [
          "inactive"
        ]
      }
    },
    "/api/wallet/link": {
      "post": {
        "deprecated": true,
        "operationId": "Wallet_link",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LinkWalletDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "501": {
            "description": "INACTIVE (P13). Gated by WALLET_ENABLED; with the flag on the service still refuses. Requires written authorization before it does anything."
          }
        },
        "tags": [
          "inactive"
        ]
      }
    },
...
  },
  "servers": [],
...
}